refactor(settings): adopt PanelBar; lift tab state out of AdvancedSettings
